api/requirements.dev.txt
fastapi[all]==0.70.1
tensorflow==2.11.0
Pillow==8.4.0
